ENGBL-1306: Rework data structures for storing subscriptions
Split the subscriptions that are for the local node and subscriptions
for a remote node out into a hash table and list respectively. This
simplifies the logic for look ups as well as improves the efficiency.

Note that due to the timing between the address of the local node being
set if is possible for subscriptions added using the remote API, but are
for the local IP, to be added as remote subscriptions. Therefore when the
local address is finally set we need to check all remote subscriptions to
confirm they are actually remote.

Change-Id: I84a9db34ee21ea99a837dfbe2de2ed6da32711d5
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>