ENGBL-1306: Implement logic for removing individual subscriptions
Add the required logic for handling the removal of a single subscription.
If the removal of the subscription results in a method with no subscribers
then we remove the method from the database. Similarly if a service is then
left with no methods we remove the service from the database.

Change-Id: I9294d97564f04d4f0f10026e98281b3fce41e1be
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>