ENGBL-1306: Implement logic for removing all subscriptions for a subscriber
Add the required logic for handling the removal of all subscriptions for a
given subscriber. If the removal of the subscription results in a method
with no subscribers then we remove the method from the database. Similarly
if a service is then left with no methods we remove the service from the
database.

Change-Id: Iac026d35f4f2783a9e854bcf63952704c95027ff
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>