ENGBL-1306: Change GList of transports to a composite client
Instead of storing a list of the transports that have subscribed to a
given notification we can create clients for each transport and add
them to a composite client. This means later on when the daemon needs
to send the published notifications there is no extra work to do, it
can simply send on the composite client.

Change-Id: I9660066d94541a7b6a738cbb3936af9041714be6
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>