ENGBL-1306: Modify publisher/subscriber tests based on refactor
Three major changes under this patch:

- publisher_subscriber_tests.c updated to make use of the new publisher
APIs that have been implemented. Note that these tests now need to start
the cmsg_pssd application for publishing to work.

- pthread_helpers_tests.c The pthread helpers for publishers have been
removed and the subscriber running in a separate thread is simply testing
whether running a cmsg_server in a separate thread works (which already has
tests). Therefore we remove these tests now.

- cmsg_pub_unit_tests.c renamed to cmsg_transport_unit_tests.c as the only
remaining functions in the file that are being tested are to do with
transports.

Change-Id: I32186c842653ee7bf0ffed4fcef7b9fbe795950a
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>