ENGBL-1306: Turn off old publisher/subscriber mechanism
Previously a cmsg publisher and subscriber would talk directly to each
other. Now after the implementation of cmsg_pssd a publisher and
subscriber do not talk directly to each other:

- The subscriber subscribes to events with cmsg_pssd.
- The publisher sends events to cmsg_pssd which forwards them to all
  interested subscribers.

Based on this we can now remove most of the existing cmsg publisher
code. The subscriber code is also changed to no longer RPC to the
publisher server (which no longer exists). A future patch will tidy up
the subscriber code.

Change-Id: I94dca93439d56245cf68053ef1b126064e8ff185
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>