ENGBL-1305: Avoid deadlock when cancelling server thread
Currently the list mutex is held when the call to 'event_server_deinit'
is made. However, the server being processed in a separate thread may
be blocked waiting for the list mutex. Subsequently it never responds
to the thread cancellation request meaning the thread holding the list
lock is blocked in pthread_join.

We avoid this problem by introducing a second mutex. The locking is required
so that multiple threads do not race creating/destroying the server. Using
the new mutex we lock the entire 'cmsg_service_listener_listen'/
'cmsg_service_listener_unlisten' functions so that the server is always
protected. Now simply release the list mutex before cancelling the server
thread.

Change-Id: I2177a6f5768abb8db0f30843065e857e3040c1b3
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>