ENGBL-1305: Associate an ID with each service listener in a process
Multiple things in a process may register for service events for the
same service. The code as it is currently written will only notify the
first listener of any existing servers. To solve this we now associate
an ID with each listen request in a process and always send the request
to cmsg_sld. When each request is received any existing servers will be
sent specifically to that listener.

Change-Id: Iac1509723b6aa4b3c97ed5e4be40c7464cba202a
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>