ALMSTK-95: Add function to connect socket with a timeout
Currently only the TIPC transport implements the ability to timeout
based on the passed in timeout value. The TIPC transport uses a
socket option for setting this timeout, the 'connect' system call
doesn't allow a timeout value to be passed in. To allow the connect
to obey a timeout value for TCP and UNIX transports we abstract the
'connect' system call in such a way that it will timeout if the connect
is not completed after a certain period of time.

Note that we do not switch the TIPC transport over to using the new
'connect_nonb' function. This is because connecting on a non-blocking
TIPC transport is not behaving similarly to TCP and for now we can simply
leave it using the previous mechanism.

Change-Id: Ie8852acc7b704b8cb1a71e14c30443a648c4f055
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>