ALMSTK-95: Default send timeout value is 5 seconds
The send on a client transport socket will hang if the socket buffer
is full. On TIPC sockets this will unblock if the link goes down, but on
TCP sockets the function can block indefinitely. Therefore ensure we only
block for a maximum of 5 seconds before giving up.

We now store the send timeout on the transport itself.

Change-Id: If40e48e1b52a6d4c08a10f93294c3c3cd1b56ea3
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>