ALMSTK-95: Default recv timeout value is 5 seconds
The recv on a client transport socket will hang if the socket buffer if a
complete CMSG packet is not in the socket receive buffer. On TIPC sockets
this will unblock if the link goes down, but on TCP sockets the function
can block indefinitely. Therefore ensure we only block for a maximum of 5
seconds before giving up.

Also fix the previous spelling mistakes of 'peek'.

Change-Id: I12704422c7cf4d16a16c2a77dafcef20c4d6e700
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>