ALMSTK-95: Default connect timeout value is 5 seconds
Default to a connect timeout value of 5 seconds. Previously connect calls
on TCP connections could hang for a long time (over a minute). Previously
the default TIPC connect timeout was 8 seconds, but since the upper link
tolerance used for VCS is 5 seconds this change should have minimal effect.

We now store the connect timeout on transport to avoid passing it around.

Change-Id: Idd63de4b1d505322a927e71c87a9b26424b671c8
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>