ALMSTK-65: Abstract accept threading onto the server structure
Returning the 'cmsg_server_accept_thread_info' structure from a number
of public library functions can cause confusion to library users. Furthermore
it isn't actually required if we just store it on the server structure.
Listening for connections in a separate thread will likely become the default
behaviour in the future and furthermore won't require a separate structure
to store the information (a global accept thread will be used). Therefore we
rework the library to store the accept thread information on the servers themselves.

Change-Id: Ib2d1f71bfb660a16837f37efa30e7a45cff5751b
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>