ALMSTK-66: Fix superficial port number endian issues
These changes are simply tidy ups and have no effect on the running
operation of the CMSG library.

- When printing the TCP port numbers in debug some places were not
converting the port number to host order.
- Change 'cmsg_service_port_get' to return a uint16_t rather than an
int as port numbers are a 16 bit unsigned integer.
- In 'cmsg_transport_write_id' don't pass the format specifier as a
pointer. This simply allows the compiler to perform more checks and
is a safer way to call printf related functions.

Change-Id: Ifab8cbfccc598133d15606d72e5f467b483d95f8
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>