ALMSTK-66: Port number must be sent as bytes to other devices
Port numbers are always stored in network byte order. Currently we
are sending this value in a uint32 field but due to the nature of
the protobuf-c encoding (all values are sent as little endian) the
value is having its byte order swapped when sending between hosts
with different byte order. Instead to ensure the value is always
received correctly we must send it as an array of bytes.

Change-Id: I02e1c275a49ebc5e1941a24499b4262aba9eee7c
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>