CR-63564: Prevent null pointer crash in cmsg_proxy_json_value_to_object
URL parameters are decoded/unescaped before attempting to parse them. In
certain cases, the result of this decoding can be a null value. If the URL
parameter is an integer type, calling strtoll on this null value causes
a crash. We now check if it is NULL and return a json_null type object in
this case. This prevents the possibility of a crash and allows an error
to be returned by json2protobuf.

Also:
* Always fall back to returning string if value is set. This will
  mean that an error is returned for booleans rather than the parameter
  being removed.
* Log an error if an unsupported conversion is performed. (shouldn't
  happen unless somebody creates a query or path parameter of an
  unsupported type.
* Only allow base 10 conversions for integers. (json number format doesn't
  support octal or hexadecimal.) string values are converted in base 10 in
  json2protobuf.
* Add unit tests for all of the above. A previous unit test
  test_cmsg_proxy_json_value_to_object__invalid_boolean is removed. The
  CR that this was added for still behaves correctly after the change.

Change-Id: Iaf7ed66bc6a4acd1ae8c5f6cf547d12fbb5468b8
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>