ALMSTK-106: Add service support check extension
This extension allows a service to be marked as supported/unsupported
at runtime. This is useful if a service has been compiled and included
in a build but no server ever implements the service at runtime. The
service being supported is specified by a specific file existing in the
file system. Any client that wishes to call the given service will check
for this file before attempting to send to the server. If the file does
not exist then ANT_CODE_UNIMPLEMENTED is returned with an optional message.

Change-Id: I9014154e9ba7bd258322366ff5554f0b10359d16
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>