CR-63662: Adding a subscription must be synchronous
When a subscriber adds a subscription they expect that once the
function that adds the subscription returns they will receive every
event that is published for the subscription they added.

Currently the act of adding a subscription is asynchronous in that
the subscriber simply sends the subscription request message and
continues immediately. Eventually the cmsg_psd process reads this
message once the glib event loop runs and notices the data on the
socket.

A problem occurs when something publishes data at the same time that
something subscribes for that data. Depending on the specific file
descriptor numbers and how the glib event loop functionality behaves
the subscription may be added before OR after the data is published
meaning the subscriber misses the event it believes it should have
received.

The fix for this is to make the subscription requests synchronous in
that the caller will not return until the subscription is fully
registered with the cmsg_psd process. Subscription requests being
synchronous was the behaviour before the publisher/subscriber functionality
was refactored so this simply brings back the old behaviour.

Note that the act of publishing data is left to be asynchronous.

Change-Id: I6ef0d2e2d7ac643c8ff98fd11ea77ccb8dd74013
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>