CR-63787: Don't publish messages if there are no subscribers
This is part 1 of 2 patches to stop publishers sending messages to
cmsg_psd when there are no subscribers for the method they are publishing
for. Sending a large number of messages to cmsg_psd for the daemon to
simply drop them when there are no subscribers can be very inefficient.

This patch changes the publisher to now register itself with cmsg_psd when
it is first created. The cmsg_psd daemon will then return a list of methods
that currently have at least one subscriber. The publisher can then store
these and then check this list before actually sending a published method
to cmsg_psd. This ensures publishers will no longer send messages to cmsg_psd
for methods with no subscribers.

The second patch will account for the fact that subscribers for a method
can be added/removed after the publisher is initialised.

Change-Id: Ib3d683d76f68a9d87040f961bb6a25115351e181
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>