CR-63787: Update publishers when methods gain/lose subscribers
This is part 2 of 2 patches to stop publishers sending messages to
cmsg_psd when there are no subscribers for the method they are publishing
for.

This patch extends the first by adding a server to the publisher, running
in a separate thread, that receives updates as subscribers come and go
for methods the publisher publishes events for. When registering with the
cmsg_psd process the publisher sends the transport information about this
server so that cmsg_psd can send updates to the publisher as required.
This change ensures that subscriptions that occur after a publisher has
been created are correctly synchronised to the publisher. Note that a
publisher only cares if there is at least one subscriber for a given
method.

Change-Id: Id8a1db5038adb86c84eb4f20bf4720694028f545
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>