CR-63787: Store the full information about all subscribers on the publisher
The previous patch changed the CMSG publisher to keep track of whether
there was a subscriber for each of the methods it publishes on. This is
now extended to store the information about every single subscriber on
the publisher. We store this in a hash table and as part of a CMSG
composite client. This will allow the events to be published directly
from the publisher to subscribers which will be done in the next patch.

Change-Id: I5fc8e5f3a3f54586c14a35396d0a277fa1e97a80
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>