CR-63787: Fix publisher subscriber functional tests
These tests are failing occasionally when running on Jenkins. Specifically
the tests are timing out which suggests the subscriber is never receiving
the published message. The likely reason for this is that the publisher
is destroyed before the send thread of the publisher has a chance to send
the message, meaning the subscriber never receives the packet. Modify the
test to ensure the subscriber receives the message before we destroy the
publisher.

Change-Id: Ia1b4efcaf7fb16e412cf1d700fa221a7d2c40566
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>