CR-63682: Use different return codes for cmsg_transport_peek_for_header
The 'cmsg_status_code' values are part of the CMSG protocol (they are
in the METHOD_REPLY header). A future patch is going to add a new
return code for 'cmsg_transport_peek_for_header' so rather than adding
a new status code we change the function to return different values and
convert to status codes as required.

Change-Id: I30fd822c70c9ba4b0b35a42b55d578279acd02a3
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>