CR-63682: Ignore spurious wake ups on CMSG server sockets
When using select/poll/epoll on sockets the kernel may spuriously
wake up the caller saying the socket is ready to read when in fact
reading the socket will block. Previously the server receive code
would log an error if it tried to peek the header on a socket that
had no data (i.e. a spurious wake up). Now the code explicitly checks
for this case and simply returns an early success when trying to
receive on a server socket with no data.

Change-Id: Ic0b88e868f5bde32404adfd35470d8aede194cd0
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>