CR-65008: Ensure usages of 'send' handle being interrupted
Under high load we are seeing the send call returning a value less
than the total requested bytes to send. This suggests the call has
been interrupted after at least 1 byte has already been sent on the
socket. Abstract the usage of the send function to now account for
this and ensure that all of the requested data is in fact sent, even
if the call is interrupted.

Note that we now also use TEMP_FAILURE_RETRY for the case that the
call is interrupted before any data is sent, rather than rely on the
library user using SA_RESTART with their signal handlers.

Change-Id: I998996602d298e667ab21f8ea44822ff8830bae0
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>