OFFPAC-62: Ensure default value is generated correctly with no package
When an enum in a .proto file with no package was used as a default
value, for example ANT_CODE_OK, it was generated with a leading
underscore, i.e. _ANT_CODE_OK. Fix this so that the generation is
correct.

On the 'ant_result' message we now explicitly set the default code
value so that when the initialiser is generated it initialises the
field with ANT_CODE_OK instead of 0. This fixes warnings when
compiling with a C++ compiler (which dont like converting integer
to enum values).

Change-Id: I8d5c650ed2ca87b23e90c36f3c67c2f9cc8076f1
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>