CR-65900: Detect when there is data but not enough for a cmsg header
Instead of return code that is converted to success then re-selecting
the socket, Allow it to timeout and return an error message if the
socket never gets enough data to read a header. This avoids a indefinite
busy loop going between the select showing data then the read returning
as there is not enough to avoid blocking.

Change-Id: I8a174ca514183beee31a0047b0734779a8c98897
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>