ENGBL-1386: Generate C protobuf types using upstream protoc-gen-c
Use the protoc-c plugin from the protobuf-c package to generate all
of the protobuf message types instead of the copied (and modified)
code in the CMSG plugin. Note that this causes the generated types
to have double underscores (and other oddities) that we previously
changed in the code we copied into the CMSG plugin. The next patch
ensures that the previous behaviour (i.e. removed double underscores)
is maintained.

Change-Id: I3874945289e6e0d9821c8465514c34c7b37bfa6c
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>