ENGBL-1386: Modify CMSG plugin to work now that protoc-c is used for types
The protoc-c plugin generates the protobuf types with double underscores
(and other oddities). Previously this was not an issue as we had copied
the protoc-c plugin code into the CMSG plugin and modified the types to
have the naming structure that we wanted. Now that we use the unmodified
protoc-c plugin for types the CMSG plugin generates typedefs and #defines
for the generated protoc-c types so that our desired naming structure is
maintained.

Note that there are now two helper files: c_helpers.cc and c_helpers_cmsg.cc.
The c_helpers.cc file is a copy of the identically named source file for the
protoc-c plugin. The c_helpers_cmsg.cc is the same file but with the modifications
for the desired CMSG naming style. The protoc-c functions are called from the "c"
namespace while the CMSG ones are called from the "cmsg" namespace. This makes it
simple to see how the naming style is converted. This is likely to be further
simplified in the future.

Change-Id: I80c52d5261b16ecf0bb975d02e1b3abb85142ca6
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>