ENGBL-1386: Remove unused code in CMSG plugin
This code was previously copied from the protoc-c plugin from the
protobuf-c package. Now we use the protoc-c plugin directly and
generate typedefs/#defines in the CMSG plugin to ensure our desired
naming can be used for the types. As such none of this code is required.
Note that future patches are likely to simplify the CMSG plugin further.

Change-Id: I6162121716e778bbd4ec0c7d36f612f0d3f82bb1
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>