TORUSAPI-258: Add new utility function to get string from an enum value
It is useful to get the enum name defined in .proto file for a specific
enum value instead of defining new strings hard-coded in applications.

The value-to-name mapping is generated by protoc, so simply find the name
via the corresponding enum descriptor, which can be accessed by using
CMSG_MSG_DESCRIPTOR(enum_type).

Note that there is no way to check during compile-time whether the enum
value (integer) is defined in the specified enum type or not.
So it's up to the caller to make sure to pass correct descriptor.

Also added some unittests.

Change-Id: If55469c907896acafbcf700eeba0c2b4aedeca6e
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>