CR-67714: Add extension for returning message serialised to file
Currently the CMSG server must always be available to respond to CMSG
api requests. This is an issue when we wish to stop processes running
but still respond to requests (at least enough to return the current
configuration). Since CMSG based daemons serialise their configuration
message to the file system (as a means of backup) we now implement the
functionality to return this serialised message directly from the
file system, avoiding the RPC to the server.

Change-Id: Id0216f23675051bc7c7b6f7877d37adc8138c86a
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>