CR-67714: Update supported_service extension to specify returned ant_code
Previously this extension was only used to indicate that a service
was unsupported on the device. Now we also use it to indicate that
a service is not currently running, but could be enabled if desired.
Therefore we now need to return ANT_CODE_FAILED_PRECONDITION or
ANT_CODE_UNIMPLEMENTED depending on the situation.

Change-Id: If8355be03b9c812b13f3fbfa5431b24c348afd25
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>