CR-66444: Catch process events in sld
If a process exits abnormally (i.e. crashes) then the information about
the service is never cleaned up. Now we listen for events via netlink
that notify us when a process crashes. By storing the pid of the process
when we add a server we can easily remove it if the process crashes.

To Fix this CR, there are separate patchs to:
1. provide functions to remove a listener server with a specfic process id
2. remove a subscriber after received notification from sld

Change-Id: I928b6edcd71cac8a98575c43050eb27c2a7fdca2
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>