CR-68267: Fix a core dump in cmsg psd
When a service_data_entry is created, g_io_add_watch() is called to add
an IOchannel into the main loop, but when a service_data_entry is freed,
the event source is not removed. This may result in a case that the
_sl_event_process callback run with a freed data pointer, hence causes
a bus error.

Also in function _sl_event_handler() of file data.c, if pointer info is
null, it should not be dereferenced.

Change-Id: I8e299a3b8bba1350e10db9426f245ede1dffeb1f
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>