CR-68267: Fix use after free in cmsg_psd
The original fix for this issue was not fully correct. The problem is
that there are two code paths for freeing the service entry (which is
responsible for stopping the service listening for that service). The
first path occurs if the subscriber for the service is removed normally
then it sends a message to cmsg_psd and if it is the only subscriber
for that service (with no publishers now) then the service entry is
removed and the service listening stopped with no issue.

The second code path occurs when cmsg_psd learns of the subscriber leaving
via the service listening functionality, before the message directly from
the subscriber is received. In this case we eventually remove the service
entry, and stop the service listening while in the glib callback function
used for the service listening. Currently we return G_SOURCE_CONTINUE
unconditionally from this callback function meaning that the source is
never removed and the reference count on the channel never reaches zero.
It is then possible for the callback function to execute again, now with
a pointer to the cmsg_sl_info that has now been freed.

The fix for this issue is to return G_SOURCE_REMOVE from the callback
function if the service entry was removed and we are no longer registered
for service listener events. Otherwise we return G_SOURCE_CONTINUE to
continue receiving the events.

To make it possible to workout if we have removed the service entry for the
given service we now improve how a subscriber is removed from the cmsg_psd
data. Instead of iterating through every service we send the service name with
the subscriber information to ensure we check only the single service the
subscriber was for.

Also for correctness sake, remove the source before shutting down the channel.
This is done simply because this is the reverse order of how they are created.

Change-Id: If7a12ee6ad8e29644c7f1509c7089f106f967efa
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>