CR-68318: Close CMSG Client on error
If an error occurs when sending to a CMSG Client the client connection
needs to be properly closed, so the socket is shutdown and removed, and
the CMSG client connection state correctly updated.  This allows further
attempts to connect or send on the socket to do the right thing rather
than using a socket that may not be valid.

Change-Id: I9318a89073b5c4110926b83306ecb060c32de01f
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>