CR-69427: Remove service listener code using linux connector
Currently the service listener daemon uses the connector functionality
to listen for processes crashing. This allows the service listener to
tidy up and registrations for the process that has crashed. Unfortunately
the connector is not supported in Linux containers and the upstream
maintainers state that they do not wish to support it. Therefore we remove
this code and will solve the issue of catching crashed processes differently.

Change-Id: I056476bd56ff2ba81a9bef753056e925ff0be3e2
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>