CR-69427: Periodically check for registrations from dead processes
If a process crashes that has registered data with the service listener
daemon then this data is never removed. To ensure this data is eventually
removed we now poll every 30 seconds for the PID in the /proc filesystem
(note that inotify does not work on /proc). This ensures that the
system will eventually correct itself and stop printing error logs
about dead connections. Since process crashes are unlikely the possibility
of error logs for up to 30 seconds is acceptable.

Change-Id: Iabc3acb2503f44b58eecdec1020c0a2d32d68e71
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>