MAINT: Add function to process messages queued by server
The existing server side queuing functionality is coupled to how it
is currently used in EXFX. Rather than refactoring it at this stage
add a new function that can be used to process the server queue no
matter the current queuing configuration on the server.

Functional tests are added for this new functionality, these tests
are currently a copy of the client side queuing tests.

Change-Id: I600df476d7cfa4ffddbace1173261fcb342d0ba8
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>