LU2019-908: Support TIPC servers in service listener
Currently only TCP and Unix socket servers are supported in the CMSG
Service Listener Daemon. This patch adds support for TIPC servers to
the daemon.
This allows a service listener to discover newly started or stopped
CMSG servers that use a TIPC transport.
Note that the local TIPC instance id is inferred from the local stack
node IP address's last byte.
The debug output on SIGUSR1 is updated to include the TIPC server
information and the TIPC related test cases are updated to ensure 100%
test pass.

Change-Id: I45bb951ac8991a28f06c1350c81105f3e261b487
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>