MAINT: Change g_list_append to prepend in the service listener
g_list_append appends to the end of the list, however the g_list only
has a pointer to the head of the list and so to get to the end
the append must iterate to the end of the list.

Our usage of the list is unordered and so it is much more
efficient to use g_list_prepend which only has to add the list item
to the start of the list.

Change-Id: Ib28df9a6838392286387468e02fc70d35713cbe2
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Hamish Martin <hamish.martin@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>