MAINT: Change g_list_append to prepend in the pub/subscriber
g_list_append appends to the end of the list, however the g_list only
has a pointer to the head of the list and so to get to the end
the append must iterate to the end of the list.

Our usage of the lists is unordered and so it is much more
efficient to use g_list_prepend which only has to add the list item
to the start of the list.

Change-Id: I5560d49f30a4da7308fe5c914d975ea00297980d
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Hamish Martin <hamish.martin@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>