AMFPI-139: Handle deep infinite recursion when traversing messages
Currently the check to avoid infinite recursion handles infinite
recursion at a single depth, i.e. A->B->A, but it does not handle
infinite recursion at anything deeper, i.e. A->B->C->A. This has
now been exposed by bringing in more 3rd party .proto files.

The easiest and correct fix is to use a set to keep track of the
descriptors we have traversed and if we have already traversed the
given descriptor then to simply break the recursion.

Change-Id: Ib2dff44354e37347bb9beff832f58fc51542c382
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>