CR-68061: Use pidfd functionality to catch processes that exit
If a process crashes that has registered data with the service listener
daemon then this data is never removed. Previously we used to detect
this by polling each process by sending a signal to it every 30 seconds.
Now the linux kernel provides 'pidfd' functionality such that we can be
notified once the given process exits - avoiding the need to poll.

Change-Id: I1250b9cdb71ddbc660459f2adff1232b40443c65
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>