MAINT: Remove duplicated boilerplate from function test code
Define common functions for created the client and server used in the
functional test suite. This allows for a significant amount of duplicated
code to be removed.

Note that the one way tests are modified slightly to now actually wait
for the server to receive the message before finishing. For now the
TIPC broadcast oneway test is disabled as the helper functions do not
handle connectionless communication. This will be fixed in the future.

Change-Id: I02ef0a34aeb90793c54f3ab49ab598d52959d76b
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>