CR-70446: Close socket on 'cmsg_server_receive' error
When a client disconnects from the server 'cmsg_server_receive' will
return an error and the socket should be closed by the server. The glib
helper, and now also the liboop helper, was not doing this. Eventually
if enough connections/disconnections were made by a client the server
could not accept anymore connections due to exceeding the number of
file descriptors for the process.

Change-Id: I8e14a645b73621bb03bc579db40a3e776990ceeb
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>