CR-70508: Don't double close sockets when processing with the glib helper
The glib documentation gives the impression the underlying file descriptor
is not closed when the channel is shutdown. However, digging into the glib
implementation it is clear that it does in fact close the file descriptor
when we call 'g_io_channel_shutdown'. Therefore the recent addition to
explicitly close the socket causes a double close of the file descriptor
which eventually leads to other problems. Remove the explicit close and
add a comment such that we shouldn't hit this issue again.

Also add a test for using the glib helper which exposed problems in how
we unref channels/sources. Specifically we were missing unref calls for
created 'GSource' structures, and we were calling unref twice for the
'GIOChannel' used to read each accepted connection.

Change-Id: I47fc1cb22d9b81e04cd7b9d3327ca3e39db72a27
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>